refactor(settings): move release-check accessors, consts, validators to settings_release_check.go (PRD #1021 M2)
